Make footnote parsing more flexible
Trying to parse out the ID was brittle because different cases use
different id structures.
