+++ app/webpacker/store/modules/footnotes_ui.js
@@ -10,14 +10,8 @@ const getters = {
 };
 
 const mutations = {
-  register: (state, payload) => {
-    let footnoteState = state.all[state.id];
-    if(footnoteState) {
-      footnoteState.push(...payload.annotationIds.filter(id => !footnoteState.includes(id)));
-    } else {
-      Vue.set(state.all, payload.id, payload.annotationIds);
-    }
-  }
+  register: (state, payload) =>
+    Vue.set(state.all, payload.id, payload.annotationIds)
 };
 
 export default {
